+++ tee-worker/client-api/parachain-api/CHANGELOG.md
@@ -7,9 +7,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+-   (#2930) `PlatformUserType`: Add `DarenMarket`
+
+### Changed
+
+-   (#2930) `PlatformUserType`: rename `MagicCraftStakingUser` to `MagicCraftStaking`
+-   (#2930) `PlatformUserType`: rename `KaratDaoUser` to `KaratDao`
+
+## [0.9.18-10] - 2024-07-15
+
+Matching version for [parachain-release v0.9.18-10](https://github.com/litentry/litentry-parachain/releases/tag/v0.9.18-10)
+
 ## [0.9.18-next.11] - 2024-07-12
 
-## Added
+### Added
 
 -   (#2877, #2889) Add `MFan` and `Mvp` to `Web3NftType`
 -   (#2884) Add `An` and `Tuna` to the list of `Web3TokenType`